requirements.txt
altgraph==0.17
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cmsmap==1.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.5
django-on-heroku==1.1.2
djangorestframework==3.12.4
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
idna==2.10
install==1.3.4
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
macholib==1.14
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.1
opencv-python==4.5.2.52
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pipenv==2020.11.15
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
psycopg2-binary==2.9.1
ptyprocess==0.7.0
py3dns==3.2.1
pycparser==2.20
Pygments==2.8.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
secure-smtplib==0.1.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.2
SpoofMAC==2.1.1
SQLAlchemy==1.4.9
sqlparse==0.4.1
terminado==0.9.3
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
twilio==6.60.0
urllib3==1.26.6
virtualenv==20.4.4
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
xlrd==2.0.1
